pyproject.toml
[project]
name = "jobtrail"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "jinja2>=3.1.4",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0.0",
    "sqlmodel>=0.0.22",
    "psycopg2-binary>=2.9.10",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
